requirements.txt
contourpy==1.0.7
customtkinter==5.1.3
cycler==0.11.0
darkdetect==0.8.0
dnspython==2.3.0
fonttools==4.39.4
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pip==23.1.2
psycopg2-binary==2.9.6
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
setuptools==65.5.1
six==1.16.0
tzdata==2023.3
wheel==0.40.0
zipp==3.15.0
cassandra-driver==3.27.0
